Move file handling, output, parsing, concurrency, debug, and report
