+++ tests/test_format.py
@@ -76,10 +76,16 @@ SOURCES = [
     "src/black/brackets.py",
     "src/black/cache.py",
     "src/black/comments.py",
+    "src/black/concurrency.py",
     "src/black/const.py",
+    "src/black/debug.py",
+    "src/black/files.py",
     "src/black/lines.py",
     "src/black/mode.py",
     "src/black/nodes.py",
+    "src/black/output.py",
+    "src/black/parsing.py",
+    "src/black/report.py",
     "src/black/rusty.py",
     "src/black/strings.py",
     "src/black/trans.py",
